fix: #122 capture: color_rgbi_run picks up INT16 samples (was all-zero)
The Color sensor in MODE 5 emits four INT16 channels per sample
(R / G / B + ambient intensity) with a typical raw range of 0..1024.
Two firmware bugs combined to capture nothing usable:

1. `apps/sensor/sensor_main.c::convert_sample()` only branched on
   `s->data_type == LUMP_DATA_INT8`.  MODE 5 publishes INT16, so the
   branch never fired and the zero-initialised record was written
   for every sample — every R/G/B/I in the .cap file came out as 0.

2. The `capture_color_rgbi_run_record_s` schema declared the four
   channels as `uint8_t`, which can only hold 0..255 even when the
   conversion is fixed.  A 1024-range value would silently truncate.

Fix
---

Schema (`apps/capture/include/capture_schema_color_rgbi_run.h`,
`apps/capture/schemas_color_rgbi_run.c`):

    uint8_t  red / green / blue / intensity   →   uint16_t each
    record_size:  8 B  →  12 B
    field tags:   u8   →  u16

`convert_sample()` now branches on the wire data type:

  - `LUMP_DATA_INT16` → read `s->data.i16[0..3]` (the real path)
  - `LUMP_DATA_INT8`  → read `s->data.i8[0..3]`  (kept as a fall-
                        back for any future packing variant)

Negative values clamp to 0 (the schema field is unsigned), and the
upper clamp is the uint16_t range.

Codegen + docs
--------------

`gen_pc_schema.py` re-run:
  - `host/CaptureViewer/src/CaptureViewer.Core/Generated/
     Schema_ColorRgbiRun.cs` — Record struct now has `ushort` fields
     at offsets 4 / 6 / 8 / 10, RecordSize = 12.
  - `docs/{ja,en}/development/capture-schemas.md` table refreshed.

Compatibility
-------------

Existing `.cap` files captured from MODE 5 against the previous
firmware are all zero anyway, so there's nothing worth keeping at
the old layout.  The schema_magic is unchanged (0x0011), so the host
parser keys off the per-file `record_size` / `field_count` header
fields — those drive the offset table the parser builds, so old
all-zero captures still open as 8-byte records with 4 u8 fields and
new captures open as 12-byte records with 4 u16 fields.  The codegen-
emitted `SchemaColorRgbiRun.Parse(Span)` is hard-coded to the new
12-byte layout; old recordings should be re-captured against this
firmware.

Verified
--------

  make                              → firmware builds clean
  dotnet test                       → 28/28 passed
  mkdocs build --strict             → ja+en clean
  hardware (operator confirmed):
    sensor color select 5 + watch   → INT16, four 0..1000 values
    sensor color capture 1500 &
    App "Send MODE CAPTURE"         → R/G/B/I traces show real data

Issue: #122

+++ apps/capture/schemas_color_rgbi_run.c
@@ -23,9 +23,9 @@ const struct capture_schema_s g_capture_schema_color_rgbi_run =
   .fields       =
   {
     CAPTURE_FIELD_INIT(color_rgbi_run, ts_us,     u32, "us",  0),
-    CAPTURE_FIELD_INIT(color_rgbi_run, red,       u8,  "raw", 0),
-    CAPTURE_FIELD_INIT(color_rgbi_run, green,     u8,  "raw", 0),
-    CAPTURE_FIELD_INIT(color_rgbi_run, blue,      u8,  "raw", 0),
-    CAPTURE_FIELD_INIT(color_rgbi_run, intensity, u8,  "raw", 0),
+    CAPTURE_FIELD_INIT(color_rgbi_run, red,       u16, "raw", 0),
+    CAPTURE_FIELD_INIT(color_rgbi_run, green,     u16, "raw", 0),
+    CAPTURE_FIELD_INIT(color_rgbi_run, blue,      u16, "raw", 0),
+    CAPTURE_FIELD_INIT(color_rgbi_run, intensity, u16, "raw", 0),
   },
 };
